# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Your API Name",
      "description": "API for managing items and user authentication.",
      "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
      "/auth/register": {
        "post": {
          "summary": "Register a new user",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User registered successfully"
            },
            "400": {
              "description": "Username already exists"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "summary": "Login a user",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful login"
            },
            "401": {
              "description": "Invalid credentials"
            }
          }
        }
      },
      "/items": {
        "post": {
          "summary": "Create an item",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Item created"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "get": {
          "summary": "Get all items",
          "responses": {
            "200": {
              "description": "A list of items"
            }
          }
        }
      },
      "/items/{item_id}": {
        "put": {
          "summary": "Update an item",
          "parameters": [
            {
              "name": "item_id",
              "in": "path",
              "required": true,
              "type": "integer"
            },
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Item updated"
            },
            "404": {
              "description": "Item not found"
            }
          }
        },
        "delete": {
          "summary": "Delete an item",
          "parameters": [
            {
              "name": "item_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Item deleted"
            },
            "404": {
              "description": "Item not found"
            }
          }
        }
      }
    }
  }
  